+++ src/Search/FilterValue.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Notion\Search;
+
+enum FilterValue: string
+{
+    case Page = "page";
+    case Database = "database";
+}