+++ src/Search/SortDirection.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Notion\Search;
+
+enum SortDirection: string
+{
+    case Ascending = "ascending";
+    case Descending = "descending";
+}