+++ src/Search/SortTimestamp.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Notion\Search;
+
+enum SortTimestamp: string
+{
+    case LastEditedTime = "last_edited_time";
+}