+++ tests/Unit/Databases/Query/ResultTest.php
@@ -12,6 +12,7 @@ class ResultTest extends TestCase
         $apiResponse = [
             "results" => [
                 [
+                    "object" => "page",
                     "id" => "a7e80c0b-a766-43c3-a9e9-21ce94595e0e",
                     "created_time" => "2020-12-08T12:00:00.000000Z",
                     "last_edited_time" => "2020-12-08T12:00:00.000000Z",